Refactor HTTP service

+++ src/services/http-service.js
@@ -11,40 +11,20 @@ const axios = Axios.create({
 });
 
 export const httpService = {
-  get(endpoint, data) {
-    return ajax(endpoint, 'GET', data);
-  },
-  post(endpoint, data) {
-    return ajax(endpoint, 'POST', data);
-  },
-  put(endpoint, data) {
-    return ajax(endpoint, 'PUT', data);
-  },
-  delete(endpoint, data) {
-    return ajax(endpoint, 'DELETE', data);
-  },
+  get: (endpoint, data) => ajax(endpoint, 'GET', data),
+  post: (endpoint, data) => ajax(endpoint, 'POST', data),
+  put: (endpoint, data) => ajax(endpoint, 'PUT', data),
+  delete: (endpoint, data) => ajax(endpoint, 'DELETE', data),
 };
 
 async function ajax(endpoint, method = 'GET', data = null) {
+  // debugger;
   try {
-    let accessToken = Cookies.get('accessToken');
+    const accessToken = await getAccessToken();
+    console.log('document.cookie:', document.cookie);
+    setAuthorizationHeader(accessToken);
 
-    if (accessToken) {
-      const decodedToken = jwtDecode(accessToken);
-
-      if (decodedToken.exp < Date.now() / 1000) {
-        //refreshing access token if close to expiry, backend will sign (and return) a new accessToken
-        const refreshedToken = await refreshAccessToken();
-        if (refreshedToken) {
-          accessToken = refreshedToken;
-        } else {
-          console.error('Failed to refresh token');
-          throw new Error('Failed to refresh token');
-        }
-      }
-      axios.defaults.headers.common.Authorization = `Bearer ${accessToken}`;
-    }
-    //Building the request
+    // Building the request and returning the response data
     return (
       await axios({ url: `${BASE_URL}${endpoint}`, method, data, params: method === 'GET' && data })
     ).data;
@@ -55,11 +35,43 @@ async function ajax(endpoint, method = 'GET', data = null) {
         const refreshedToken = await refreshAccessToken();
         if (refreshedToken) return ajax(endpoint, method, data);
       } catch (refreshError) {
-        sessionStorage.clear();
-        localStorage.clear();
-        window.location.assign('/');
+        console.error('Failed to refresh token', refreshError);
+        throw refreshError;
+        // sessionStorage.clear();
+        // localStorage.clear();
+        // window.location.assign('/');
       }
     }
     throw err;
+  }
+}
+
+async function getAccessToken() {
+  let accessToken = Cookies.get('accessToken');
+  if (!accessToken) return null;
+
+  const decodedToken = jwtDecode(accessToken);
+  if (decodedToken.exp < Date.now() / 1000) {
+    try {
+      const refreshedToken = await refreshAccessToken();
+      if (refreshedToken) {
+        accessToken = refreshedToken;
+      } else {
+        console.error('Failed to refresh token');
+        throw new Error('Failed to refresh token');
+      }
+    } catch (refreshError) {
+      console.error('Failed to refresh token', refreshError);
+      sessionStorage.clear();
+      localStorage.clear();
+      // window.location.assign('/');
+    }
+  }
+  return accessToken;
+}
+
+function setAuthorizationHeader(accessToken) {
+  if (accessToken) {
+    axios.defaults.headers.common.Authorization = `Bearer ${accessToken}`;
   }
 }
